Co-Locate Stories
See #32
