+++ config/storybook/config.js
@@ -1,6 +1,8 @@
 import { configure } from '@storybook/react';
 import './styles.css';
 
+const req = require.context('../../src', true, /\.stories\.tsx$/);
+
 configure(() => {
-  require('../../src/stories');
+  req.keys().forEach(filename => req(filename));
 }, module);
