+++ src/components/App/index.tsx
@@ -6,7 +6,7 @@ import Footer from '../../containers/Footer';
 import Queue from '../../containers/Queue';
 import { Albums } from '../../containers/Albums';
 import Artist from '../../containers/Artist';
-import Album from '../../containers/Album/index';
+import Album from '../../containers/Album';
 
 import { Providers } from './Providers';
 import Title from '../Title';
@@ -23,7 +23,7 @@ import {
   songs,
   withId,
   withIdFromProps,
-} from '../../paths';
+} from '../../utils/paths';
 import store from '../../redux/store';
 
 import * as styles from './styles.css';
