+++ src/components/Like/Like.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import StatefulComponent from './StatefulComponent';
-import Like from '../components/Like';
+import StatefulComponent from '../../utils/StatefulComponent';
+import Like from '.';
 
 storiesOf('Like', module).add('interactive', () => (
   <StatefulComponent state={{ like: true }}>
