+++ src/components/PlaybackButton/PlaybackButton.stories.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import StatefulComponent from './StatefulComponent';
-import PlaybackButton from '../components/PlaybackButton';
-import * as styles from './PlaybackButton.css';
+import StatefulComponent from '../../utils/StatefulComponent';
+import PlaybackButton from '.';
+import * as styles from './PlaybackButton.stories.css';
 
 storiesOf('PlaybackButton', module).add('interactive', () => (
   <StatefulComponent state={{ playing: true }}>
