+++ src/components/Slider/Slider.stories.tsx
@@ -2,8 +2,8 @@ import * as React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import StatefulComponent from './StatefulComponent';
-import Slider from '../components/Slider';
+import StatefulComponent from '../../utils/StatefulComponent';
+import Slider from '.';
 
 storiesOf('Slider', module).add('interactive', () => (
   <StatefulComponent state={{ value: 50 }}>
