+++ src/components/VolumeSlider/VolumeSlider.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import StatefulComponent from './StatefulComponent';
-import VolumeSlider from '../components/VolumeSlider';
+import StatefulComponent from '../../utils/StatefulComponent';
+import VolumeSlider from '.';
 
 storiesOf('VolumeSlider', module).add('interactive', () => (
   <StatefulComponent state={{ volume: 1 }}>
