+++ src/containers/Artist/component/index.tsx
@@ -5,7 +5,7 @@ import { pluralize } from '../../../utils';
 import { AlbumArtwork } from '../../../components/AlbumArtwork';
 import { ArtworkTwoInfo } from '../../../components/ArtworkTwoInfo';
 import { Link } from 'react-router-dom';
-import { album as albumPath } from '../../../paths';
+import { album as albumPath } from '../../../utils/paths';
 import { ArtistQuery_artist } from '../../../__generated__/ArtistQuery';
 import Title from '../../../components/Title';
