--- src/stories/index.ts
@@ -1,13 +1,0 @@
-import './Like';
-import './PlaybackButton';
-import './Slider';
-import './VolumeSlider';
-import './Collage';
-import './Artwork';
-import './PlaybackArtwork';
-import './SongList';
-import './Icons';
-import './SongCollage';
-import './Queue';
-import './keyboardInteraction';
-import './Dropdown';
